+++ tests/unit/test_output.py
@@ -113,14 +113,13 @@ def test_echo_request(mocker, capfd, monkeypatch, headers, json, expected):
 
 
 @pytest.mark.parametrize(
-    "status_code, reason, version, headers, json, text, expected",
+    "status_code, reason, version, headers, content, expected",
     [
         (
             200,
             "OK",
             11,
             {"Content-Type": "application/json"},
-            None,
             None,
             dedent(
                 """\
@@ -137,7 +136,6 @@ def test_echo_request(mocker, capfd, monkeypatch, headers, json, expected):
             10,
             {"Content-Type": "application/json"},
             None,
-            None,
             dedent(
                 """\
                     HTTP/1.0 201 Created
@@ -152,8 +150,7 @@ def test_echo_request(mocker, capfd, monkeypatch, headers, json, expected):
             "OK",
             11,
             {"Content-Type": "application/json"},
-            {"id": 1},
-            None,
+            b'{"id": 1}',
             dedent(
                 """\
                     HTTP/1.1 200 OK
@@ -174,7 +171,6 @@ def test_echo_request(mocker, capfd, monkeypatch, headers, json, expected):
             11,
             None,
             None,
-            None,
             dedent(
                 """\
                     HTTP/1.1 200 OK
@@ -188,14 +184,29 @@ def test_echo_request(mocker, capfd, monkeypatch, headers, json, expected):
             "OK",
             11,
             {"Content-Type": "text/plain"},
-            None,
-            "message",
+            b"message",
             dedent(
                 """\
                     HTTP/1.1 200 OK
                     Content-Type: text/plain
 
                     message
+
+
+                """
+            ),
+        ),
+        (
+            200,
+            "OK",
+            11,
+            None,
+            b"\0",
+            dedent(
+                """\
+                    HTTP/1.1 200 OK
+
+                    NOTE: binary data not shown
 
 
                 """
@@ -211,8 +222,7 @@ def test_echo_response(
     reason,
     version,
     headers,
-    json,
-    text,
+    content,
     expected,
 ):
     monkeypatch.setenv("ANYMOTION_STDOUT_ISSHOW", "True")
@@ -222,8 +232,7 @@ def test_echo_response(
     response_mock.return_value.reason = reason
     response_mock.return_value.raw.version = version
     response_mock.return_value.headers = headers
-    response_mock.return_value.json.return_value = json
-    response_mock.return_value.text = text
+    response_mock.return_value.content = content
 
     echo_response(response_mock())